Move wrapper submodule to root
